Feat: Implementação do Controller para Person

+++ back-end/src/main/java/com/pms/calprop/dto/PersonResponse.java
@@ -2,6 +2,6 @@ package com.pms.calprop.dto;
 
 import java.math.BigDecimal;
 
-public record PersonResponse(Long id, String name, BigDecimal salary, BigDecimal reservePercentage) {
+public record PersonResponse(Long id, String name, BigDecimal salary, Double reservePercentage) {
 
 }
